Fix to invalidate browse cache on control page.
Bug: Browser caches the control page on first GET request. When user modifies
the content of the page, the modifications are done with AJAX requests
which are not part of the GET requested page in the browser cahe. When user
returns to the page with browser's back button, browser will only render the
page that is on the cache.

Solution: This fix sets an pre render event for the view to set a HTTP
header to tell the browser to not to cache the content of the page.

Info:
http://stackoverflow.com/questions/14101380/force-jsf-to-refresh-page-view-form-when-opened-via-link-or-back-button